requirements.txt
numpy>=1.20.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.0
pdfminer.six>=20221105
tiktoken>=0.5.0
langchain>=0.1.0
pytest>=7.0.0
python-dotenv>=1.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
openai>=1.0.0
pyyaml>=6.0
